+++ src/app/api/market/[id]/route.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_LANG } from '@/const/locale';
-import { getAgentJSON } from '@/const/url';
+
+import { AgentMarket } from '../AgentMarket';
 
 export const runtime = 'edge';
 
@@ -8,11 +9,13 @@ export const GET = async (req: Request, { params }: { params: { id: string } }) 
 
   const locale = searchParams.get('locale');
 
+  const market = new AgentMarket();
+
   let res: Response;
 
-  res = await fetch(getAgentJSON(params.id, locale as any));
+  res = await fetch(market.getAgentUrl(params.id, locale as any));
   if (res.status === 404) {
-    res = await fetch(getAgentJSON(params.id, DEFAULT_LANG));
+    res = await fetch(market.getAgentUrl(params.id, DEFAULT_LANG));
   }
 
   return res;