+++ src/features/Conversation/ChatList/Extras/type.ts
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+import { ChatMessage } from '@/types/message';
+
+export type RenderMessageExtra = FC<ChatMessage>;