+++ src/features/Conversation/ChatList/Messages/Assistant.tsx
@@ -1,0 +1,29 @@
+import { ReactNode, memo } from 'react';
+
+import { useChatStore } from '@/store/chat';
+import { chatSelectors } from '@/store/chat/selectors';
+import { ChatMessage } from '@/types/message';
+import { isFunctionMessageAtStart } from '@/utils/message';
+
+import Inspector from '../Plugins/Inspector';
+import { DefaultMessage } from './Default';
+
+/**
+ * 用于渲染机器人发出的消息
+ */
+export const AssistantMessage = memo<
+  ChatMessage & {
+    editableContent: ReactNode;
+  }
+>(({ id, plugin, content, ...props }) => {
+  const fcProps = useChatStore(chatSelectors.getFunctionMessageProps({ content, id, plugin }));
+
+  if (!isFunctionMessageAtStart(content))
+    return <DefaultMessage content={content} id={id} {...props} />;
+
+  return (
+    <div id={id}>
+      <Inspector {...fcProps} />
+    </div>
+  );
+});