+++ src/store/global/helpers.ts
@@ -1,4 +1,4 @@
-import { settingsSelectors } from './selectors';
+import { settingsSelectors } from './slices/settings/selectors';
 import { useGlobalStore } from './store';
 
 const getCurrentLanguage = () => settingsSelectors.currentLanguage(useGlobalStore.getState());