+++ src/store/tool/helpers.ts
@@ -1,14 +1,15 @@
-import { LobeChatPluginManifest, PluginSchema } from '@lobehub/chat-plugin-sdk';
+import { PluginSchema } from '@lobehub/chat-plugin-sdk';
 
+import { MetaData } from '@/types/meta';
 import { LobeTool } from '@/types/tool';
 
 const getPluginFormList = (list: LobeTool[], id: string) => list?.find((p) => p.identifier === id);
 
-const getPluginTitle = (meta?: LobeChatPluginManifest['meta']) => meta?.title;
-const getPluginDesc = (meta?: LobeChatPluginManifest['meta']) => meta?.description;
+const getPluginTitle = (meta?: MetaData) => meta?.title;
+const getPluginDesc = (meta?: MetaData) => meta?.description;
 
-const getPluginTags = (meta?: LobeChatPluginManifest['meta']) => meta?.tags;
-const getPluginAvatar = (meta?: LobeChatPluginManifest['meta']) => meta?.avatar || '🧩';
+const getPluginTags = (meta?: MetaData) => meta?.tags;
+const getPluginAvatar = (meta?: MetaData) => meta?.avatar || '🧩';
 
 const isCustomPlugin = (id: string, pluginList: LobeTool[]) =>
   pluginList.some((i) => i.identifier === id && i.type === 'customPlugin');