+++ src/types/settings.ts
@@ -5,37 +5,8 @@ import { LocaleMode } from '@/types/locale';
 import type { LobeAgentSession } from '@/types/session';
 
 export interface GlobalBaseSettings {
-  /**
-   * @deprecated
-   */
-  OPENAI_API_KEY?: string;
-
   avatar: string;
-  /**
-   * @deprecated
-   */
-  compressThreshold?: number;
-  /**
-   * @deprecated
-   */
-  enableCompressThreshold?: boolean;
-  /**
-   * @deprecated
-   */
-  enableHistoryCount?: boolean;
-  /**
-   * @deprecated
-   */
-  enableMaxTokens?: boolean;
-  /**
-   * @deprecated
-   */
-  endpoint?: string;
   fontSize: number;
-  /**
-   * @deprecated
-   */
-  historyCount?: number;
   language: LocaleMode;
   neutralColor?: NeutralColors;
   password: string;
@@ -75,13 +46,23 @@ export interface GlobalTTSConfig {
 
 export type LLMBrand = keyof GlobalLLMConfig;
 
+export interface GlobalTool {
+  dalle: {
+    autoGenerate: boolean;
+  };
+}
 /**
  * 配置设置
  */
 export interface GlobalSettings extends GlobalBaseSettings {
   defaultAgent: GlobalDefaultAgent;
   languageModel: GlobalLLMConfig;
+  tool: GlobalTool;
   tts: GlobalTTSConfig;
 }
 
 export type ConfigKeys = keyof GlobalSettings;
+
+export interface GlobalServerConfig {
+  customModelName?: string;
+}